Workaround test ordering issue by just merging them
