+++ opencensus-shim/src/test/java/io/opentelemetry/opencensusshim/InteroperabilityTest.java
@@ -17,6 +17,20 @@ import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import io.opencensus.common.Duration;
+import io.opencensus.stats.Aggregation;
+import io.opencensus.stats.BucketBoundaries;
+import io.opencensus.stats.Measure;
+import io.opencensus.stats.Stats;
+import io.opencensus.stats.StatsRecorder;
+import io.opencensus.stats.View;
+import io.opencensus.stats.ViewManager;
+import io.opencensus.tags.TagContext;
+import io.opencensus.tags.TagKey;
+import io.opencensus.tags.TagMetadata;
+import io.opencensus.tags.TagValue;
+import io.opencensus.tags.Tagger;
+import io.opencensus.tags.Tags;
 import io.opencensus.trace.Annotation;
 import io.opencensus.trace.AttributeValue;
 import io.opencensus.trace.Link;
@@ -36,6 +50,11 @@ import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
 import io.opentelemetry.sdk.common.CompletableResultCode;
+import io.opentelemetry.sdk.metrics.data.DoublePointData;
+import io.opentelemetry.sdk.metrics.data.LongPointData;
+import io.opentelemetry.sdk.metrics.data.MetricData;
+import io.opentelemetry.sdk.metrics.data.MetricDataType;
+import io.opentelemetry.sdk.metrics.data.PointData;
 import io.opentelemetry.sdk.trace.SdkTracerProvider;
 import io.opentelemetry.sdk.trace.SpanProcessor;
 import io.opentelemetry.sdk.trace.data.SpanData;
@@ -43,7 +62,10 @@ import io.opentelemetry.sdk.trace.data.StatusData;
 import io.opentelemetry.sdk.trace.export.SimpleSpanProcessor;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
 import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -55,7 +77,7 @@ import org.mockito.quality.Strictness;
 
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
-class TraceInteroperabilityTest {
+class InteroperabilityTest {
 
   private static final String NULL_SPAN_ID = "0000000000000000";
 
@@ -333,6 +355,151 @@ class TraceInteroperabilityTest {
     }
     Tracing.getExportComponent().shutdown();
     verify(spanExporter, never()).export(anyCollection());
+  }
+
+  @Test
+  @SuppressWarnings("deprecation") // Summary is deprecated in census
+  void testSupportedMetricsExportedCorrectly() {
+    Tagger tagger = Tags.getTagger();
+    Measure.MeasureLong latency =
+        Measure.MeasureLong.create("task_latency", "The task latency in milliseconds", "ms");
+    Measure.MeasureDouble latency2 =
+        Measure.MeasureDouble.create("task_latency_2", "The task latency in milliseconds 2", "ms");
+    StatsRecorder statsRecorder = Stats.getStatsRecorder();
+    TagKey tagKey = TagKey.create("tagKey");
+    TagValue tagValue = TagValue.create("tagValue");
+    View longSumView =
+        View.create(
+            View.Name.create("long_sum"),
+            "long sum",
+            latency,
+            Aggregation.Sum.create(),
+            ImmutableList.of(tagKey));
+    View longGaugeView =
+        View.create(
+            View.Name.create("long_gauge"),
+            "long gauge",
+            latency,
+            Aggregation.LastValue.create(),
+            ImmutableList.of(tagKey));
+    View doubleSumView =
+        View.create(
+            View.Name.create("double_sum"),
+            "double sum",
+            latency2,
+            Aggregation.Sum.create(),
+            ImmutableList.of());
+    View doubleGaugeView =
+        View.create(
+            View.Name.create("double_gauge"),
+            "double gauge",
+            latency2,
+            Aggregation.LastValue.create(),
+            ImmutableList.of());
+    ViewManager viewManager = Stats.getViewManager();
+    viewManager.registerView(longSumView);
+    viewManager.registerView(longGaugeView);
+    viewManager.registerView(doubleSumView);
+    viewManager.registerView(doubleGaugeView);
+    FakeMetricExporter metricExporter = new FakeMetricExporter();
+    OpenTelemetryMetricsExporter.createAndRegister(metricExporter, Duration.create(0, 5000));
+
+    TagContext tagContext =
+        tagger
+            .emptyBuilder()
+            .put(tagKey, tagValue, TagMetadata.create(TagMetadata.TagTtl.UNLIMITED_PROPAGATION))
+            .build();
+    try (io.opencensus.common.Scope ss = tagger.withTagContext(tagContext)) {
+      statsRecorder.newMeasureMap().put(latency, 50).record();
+      statsRecorder.newMeasureMap().put(latency2, 60).record();
+    }
+    List<List<MetricData>> exported = metricExporter.waitForNumberOfExports(3);
+    List<MetricData> metricData =
+        exported.get(2).stream()
+            .sorted(Comparator.comparing(MetricData::getName))
+            .collect(Collectors.toList());
+    assertThat(metricData.size()).isEqualTo(4);
+
+    MetricData metric = metricData.get(0);
+    assertThat(metric.getName()).isEqualTo("double_gauge");
+    assertThat(metric.getDescription()).isEqualTo("double gauge");
+    assertThat(metric.getUnit()).isEqualTo("ms");
+    assertThat(metric.getType()).isEqualTo(MetricDataType.DOUBLE_GAUGE);
+    assertThat(metric.getDoubleGaugeData().getPoints().size()).isEqualTo(1);
+    PointData pointData = metric.getDoubleGaugeData().getPoints().iterator().next();
+    assertThat(((DoublePointData) pointData).getValue()).isEqualTo(60);
+    assertThat(pointData.getLabels().size()).isEqualTo(0);
+
+    metric = metricData.get(1);
+    assertThat(metric.getName()).isEqualTo("double_sum");
+    assertThat(metric.getDescription()).isEqualTo("double sum");
+    assertThat(metric.getUnit()).isEqualTo("ms");
+    assertThat(metric.getType()).isEqualTo(MetricDataType.DOUBLE_SUM);
+    assertThat(metric.getDoubleSumData().getPoints().size()).isEqualTo(1);
+    pointData = metric.getDoubleSumData().getPoints().iterator().next();
+    assertThat(((DoublePointData) pointData).getValue()).isEqualTo(60);
+    assertThat(pointData.getLabels().size()).isEqualTo(0);
+
+    metric = metricData.get(2);
+    assertThat(metric.getName()).isEqualTo("long_gauge");
+    assertThat(metric.getDescription()).isEqualTo("long gauge");
+    assertThat(metric.getUnit()).isEqualTo("ms");
+    assertThat(metric.getType()).isEqualTo(MetricDataType.LONG_GAUGE);
+    assertThat(metric.getLongGaugeData().getPoints().size()).isEqualTo(1);
+
+    metric = metricData.get(3);
+    assertThat(metric.getName()).isEqualTo("long_sum");
+    assertThat(metric.getDescription()).isEqualTo("long sum");
+    assertThat(metric.getUnit()).isEqualTo("ms");
+    assertThat(metric.getType()).isEqualTo(MetricDataType.LONG_SUM);
+    assertThat(metric.getLongSumData().getPoints().size()).isEqualTo(1);
+    pointData = metric.getLongSumData().getPoints().iterator().next();
+    assertThat(((LongPointData) pointData).getValue()).isEqualTo(50);
+    assertThat(pointData.getLabels().size()).isEqualTo(1);
+    assertThat(pointData.getLabels().get(tagKey.getName())).isEqualTo(tagValue.asString());
+  }
+
+  @Test
+  void testUnsupportedMetricsDoesNotGetExported() throws InterruptedException {
+    Tagger tagger = Tags.getTagger();
+    Measure.MeasureLong latency =
+        Measure.MeasureLong
+            .create("task_latency_distribution", "The task latency in milliseconds", "ms");
+    StatsRecorder statsRecorder = Stats.getStatsRecorder();
+    TagKey tagKey = TagKey.create("tagKey");
+    TagValue tagValue = TagValue.create("tagValue");
+    View view =
+        View.create(
+            View.Name.create("task_latency_distribution"),
+            "The distribution of the task latencies.",
+            latency,
+            Aggregation.Distribution.create(
+                BucketBoundaries.create(ImmutableList.of(100.0, 150.0, 200.0))),
+            ImmutableList.of(tagKey));
+    ViewManager viewManager = Stats.getViewManager();
+    viewManager.registerView(view);
+    FakeMetricExporter metricExporter = new FakeMetricExporter();
+    OpenTelemetryMetricsExporter.createAndRegister(metricExporter, Duration.create(0, 500));
+
+    TagContext tagContext =
+        tagger
+            .emptyBuilder()
+            .put(tagKey, tagValue, TagMetadata.create(TagMetadata.TagTtl.UNLIMITED_PROPAGATION))
+            .build();
+    try (io.opencensus.common.Scope ss = tagger.withTagContext(tagContext)) {
+      statsRecorder.newMeasureMap().put(latency, 50).record();
+    }
+    // Sleep so that there is time for export() to be called.
+    Thread.sleep(2);
+    // This is 0 in case this test gets run first, or by itself.
+    // If other views have already been registered in other tests, they will produce metric data, so
+    // we are testing for the absence of this particular view's metric data.
+    List<List<MetricData>> allExports = metricExporter.waitForNumberOfExports(0);
+    if (!allExports.isEmpty()) {
+      for (MetricData metricData : allExports.get(allExports.size() - 1)) {
+        assertThat(metricData.getName()).isNotEqualTo("task_latency_distribution");
+      }
+    }
   }
 
   private static void createOpenCensusScopedSpanWithChildSpan(
